Moved AnkiBindingConfig into the Application Layer.
